Initial project setup
Summary:
This is an initial project setup for C-API and C# binding.

I am seeking feedback primarily on the folder structure. The concrete C-API interface and tests are just placeholders

NOTE: The bindings will be exported to GitHub eventually, but not until they are stabilized. D59236081 is to skip syncing for the bindings, which should be landed before this Diff.

Reviewed By: yutingye, nickyhe-gemini

Differential Revision: D59132450

fbshipit-source-id: 8740f4c0d294ad5b23fb0e3ee5168120abf8c42e