Update rerun log mesh to take an optional color input
Summary:
momentum::logMesh() currently checks whether the character has mesh colors, and uses the character's color for rendering. However, our characters always have a default value of black, so we are rendering black characters by default. This diff adds an option to provide an input color to render the character mesh.

Sneak in a change to reduce marker sizes.

Reviewed By: nickyhe-gemini

Differential Revision: D60421838

fbshipit-source-id: d3d6eb7094e5950d1b11173b2e88ae1fcc8ae790