Add calibration options to refine_motion
Summary:
Previously, refine_motion is meant to be a post-process to add smoothing and reduce collision to an existing motion. So it only solves for the poses and keep all the calibration parameters fixed.

This diff adds options to calibrate either or both identity parameters and locator offsets. We can use it to solve all the parameters if we have a good starting point, eg. initialization from a neural network prediction.

PyMomentum binding is updated accordingly.

Reviewed By: yc-fb

Differential Revision: D59874379

fbshipit-source-id: 75a2bc1fa86948706b064f233ec05292f656975e