Safer enum conversion for FBX coordinate system
Summary:
Currently, `static_cast` is used to convert from momentum enums to FBX's enums, assuming the enum values are identical. However, this has a potential issue in that it may not work silently if FBX changes the assigned enum values while keeping the enum names the same.

To avoid this potential issue, this change converts the enums explicitly by relying on the enum names.

Reviewed By: juliencbmeta

Differential Revision: D60801206

fbshipit-source-id: d7876ae671166df3afcfdd1d49e235ff2d19f7f5