Expose function to get joined parameters from frame parameters
Summary:
Sequence solver doesn't understand vector of model parameters but it understands a joined parameters vector where a single vector has parameters of all frames.

So, to call sequence solver's solve, we need to pass that joined parameters vector.

Exposing function of sequence solver which allows converting vector of model parameteres to joined parameters. This helps in getting the joined parameters vector which can be passed to the sequence solver.

Reviewed By: cdtwigg, jeongseok-meta

Differential Revision: D60846314

fbshipit-source-id: 81b4c2b26aec7cd39e8ee397876078e7b66a2aa2