Fix computing scale of Transform from Eigen::Affine3
Summary:
Fix the incorrect computation of scale from the scaled rotation part of a 4x4 matrix or Eigen::Affine3.

Previously, a quaternion was created from a scaled rotation and then the scale was calculated by normalizing the quaternion. However, a scaled rotation matrix leads to an invalid quaternion in the first place, so the computed scale was not correct.

Now, the scale is first computed by taking the norm of the first column of the scaled rotation matrix, assuming it's uniformly scaled (so the norm of any columns and rows should be the same, but the first column was chosen as Eigen matrices are column-major).

Regression tests have also been added for this

Reviewed By: EscapeZero

Differential Revision: D61545672

fbshipit-source-id: a4dfa15d315cb9f7c3005f49b0168e2d2dec6af4