transformPose supports making the output continuous (removing Euler flips)
Summary:
We are having problems with the transformed poses having Euler flips due to being solved independently   at each frame.  To eliminate these, we will need to solve sequentially and start from the previous frame's rigid pose result rather than starting from the source parameters at each frame.

The trade-off here is that we lose the ability to run the solves in parallel, but this should be a good trade-off in practice.  If we need to later we can make the solve lazy.

Reviewed By: jeongseok-meta

Differential Revision: D62149618

fbshipit-source-id: 5e19ef49e8754c13af8377e04dccb7e091300a42