Add CMake options to use system-installed gtest and pybind11 (#75)
Summary:
Fetching build dependencies is convenient for users as it allows them to build packages without explicitly installing them, as they will be downloaded during the CMake configuration stage. However, this approach can sometimes be an antipattern for package managers because (1) it prevents them from using updated versions of packages, as the downloading version is hardcoded (although there is a workaround by taking the version as a CMake option, but it's not common), and (2) some package managers don't allow downloading other packages during the building process due to various reasons such as security concerns.

This change provides users with the option to choose which approach to use.

Pull Request resolved: https://github.com/facebookincubator/momentum/pull/75

Test Plan:
GitHub CI

## Checklist:

- [x] Adheres to the [style guidelines](https://facebookincubator.github.io/momentum/docs/developer_guide/style_guide)
- [x] Codebase formatted by running `pixi run lint`

Reviewed By: yutingye

Differential Revision: D62703025

Pulled By: jeongseok-meta

fbshipit-source-id: 54324d347dcad085b2bca2db65844faa16649849