pymomentum: gpu_character supports joint limits.
Summary:
There are a bunch of use cases where people are using mopy to perform gradient descent to fit body poses.  One thing that appears to be missing from these use cases is any way to enforce parameter limits.  In particular the ellipsoid limit is getting ignored due to how complicated it is, but even model parameter limits are typically not used except maybe to clamp now and then.  We'd like to support a penalty-based system like momentum was designed for, here's an implementation that uses native Pytorch and should run entirely on the GPU.

This required adding introspection for the parameter limits to pymomentum, but this seems like useful functionality regardless.

Reviewed By: jeongseok-meta

Differential Revision: D62669955

fbshipit-source-id: 699b989cdbba2b8f728484b9978ce12aeabdfad1