Update PR template (#87)
Summary:
Move the `Checklist` section above `Test Plan` in the GitHub PR template, so that `Checklist` is not included in `Test Plan` when imported. This is because all sections below the `Test Plan` section in a GitHub pull request are automatically included in the `Test Plan` section when imported.

Pull Request resolved: https://github.com/facebookincubator/momentum/pull/87

Test Plan:
Check the imported PR

## Checklist:

- [x] Adheres to the [style guidelines](https://facebookincubator.github.io/momentum/docs/developer_guide/style_guide)
- [x] Codebase formatted by running `pixi run lint`

Reviewed By: juliencbmeta

Differential Revision: D63299832

Pulled By: jeongseok-meta

fbshipit-source-id: a3b3348f54ea7439c424f5744178092c3f477650