Fix for bug in SimdNormalConstraints.
Summary: We are allocating the blocks of memory for the SimdConstraints but not initializing them.  Then when we zero out the constraints we only zero out the weights.  This is supposed to ensure that constraints that aren't "activated" have zero effect on the solve.  However, if we end up with NaNs in any of the fields (like the normals) then multiplying by zero does't actually zero them out: the NaNs still win out, and this causes the solver to explode when these NaNs appear in the Jacobian.  Therefore we need to initialize the memory to all zeros at allocation time to eliminate the possibility of getting NaN, Inf, or denormals in the solve.

Reviewed By: yutingye, tangchengcheng, jeongseok-meta

Differential Revision: D64223726

fbshipit-source-id: 0453afbca28ab0bafe5c2dea8deb856ee3632b2b