Remove parameters with zero weights.
Summary: In compact_v5, we use a single line to get all the parameters in the correct order (with flex params at the end).  This is a workaround for the fact that the model file currently doesn't support pre-declaring the parameters.  However this causes performance issues in tracking because it seems like every parameter affects the root (hence you end up with dense Jacobians, etc).  Let's remove the zero-weight parameters once they've been parsed and done their job of getting the order right.

Reviewed By: tangchengcheng, jeongseok-meta

Differential Revision: D64627640

fbshipit-source-id: 80acae61725749f897c299eadea6a460855b4c5d