Add code to write out parameter limits. (#134)
Summary:

I want to be able to test whether reading and writing works correctly, and being able to round trip like this will be helpful.  

In addition it might be useful to be able to write the entire model file out in an automated fashion and this is a first step.

Differential Revision: D66413501