Additional control over simplification. (#147)
Summary:
Pull Request resolved: https://github.com/facebookincubator/momentum/pull/147

We expose the simplify() function which is a way to reduce the size of the skeleton to make FK/IK more efficient.  However, we have found that it does not provide enough granularity for what people want, so we added separate simplifySkeleton() and simplifyParameterTransform() functions to c++ momentum.  Let's expose those out through pymomentum.

Note that simplifyParameterTransform was already available through the reduce_to_selected_model_parameters() function, but IMO that naming is not ideal and it's not sufficiently discoverable.  Having simplify_skeleton and simplify_parameter_transform functions seems like a better way to go in the long run, and we should eventually deprecate the reduce_to_selected_model_parameters() function.

Also changing some of the asserts in simplifySkeleton() to exceptions, since in calling code it would not be unexpected to forget to include a parent joint in a simplified skeleton, so we should not completely kill the process.

Reviewed By: jeongseok-meta

Differential Revision: D66664436

fbshipit-source-id: 18193982c8c59b234911be8777f3b5706139fb58