Adding half-plane constraints. (#143)
Summary:

We want to structure certain shoulder constraints as half-plane constraints, that is, a*p1 + b*p2 - c >= 0.  This will allow us to cave out regions of the space where the parameters need to act together, for example, (p1 - p2 >= 0) means p1 should always be at least as big as p2.  

Note that we are not bothering to implement this in a piecewise fashion like we did linear constraints because you can always carve out an arbitrary convex volume with a set of half-plane constraints.  One possible extension might be to allow using more than 2 variables in the plane equation if desired.  

Note that here instead of trying to single out a "targetParameter" and "referenceParameter" like we use in the linear constraint I'm just using "param1" and "param2" to emphasize the symmetry between the parameters, but I could be convinced otherwise if people want something different.

Reviewed By: jeongseok-meta

Differential Revision: D66522514