Use "linear" instead of "linearJoint". (#169)
Summary:

For minmax limits, we don't require the user to say whether it's a minmaxJoint limit or a minmax limit on a model parameter.  We can use the same heuristic here and let the user say "linear" while using the actual parameter name to infer whether it's a joint or model param.

Reviewed By: jeongseok-meta

Differential Revision: D67112359