Change param transform by constructing a new character.
Summary: The previous code copied the existing character, then set the parameter transform and limits.  This could cause issues in the future if we ever do anything nontrivial in the constructor with the parameter transform/limits, most other places we construct a new Character with a full set of arguments instead.

Reviewed By: hzsydy

Differential Revision: D67920763

fbshipit-source-id: 78d960d48ea09a66270b8d39f03c8e5395481b5a