When calling simplify(), always include the root joint.
Summary: Because we don't pass any arguments to simplify() and it just calls parametersToJoints, we need it to do the right thing and include the root joint regardless of whether it's being driven.  Otherwise the simplifySkeleton call fails and this is not an expected result.  Note that this wasn't a problem before because we always had parameters driving the root but it's no longer the case for Tinity v4 and above since all transform parameters have been moved to b_root.

Reviewed By: jeongseok-meta

Differential Revision: D68297713

fbshipit-source-id: 18a80cab0d1b63012fb786066c85aaefa931532b