Generate model parameters for the URDF model's DoFs (#197)
Summary:
Pull Request resolved: https://github.com/facebookincubator/momentum/pull/197

In the D68995554, the URDF parser only loaded the skeletal hierarchy, leaving the model parameters as 'identity.' This meant that the model parameters were merely a stack of joint parameters (7 DoFs for translation, rotation, and scale), without considering the actual joint types.

This Diffs is to generate model parameters that correspond to the degrees of freedom specific to each joint type. For example, a revolute joint is assigned a single model parameter that allows rotation along a specified axis in the URDF. Consequently, the number of model parameters matches the total DoFs of the URDF model, with an additional 6 DoFs for the root if it's a free-floating model.

Please refer to the in-code documentation for more details.

Reviewed By: cdtwigg

Differential Revision: D69249256

fbshipit-source-id: 76093c03d22c79410fb5078431014945aadf4ec2