Convert quaternion library to python.
Summary:
When I wrote the quaternion library, I was thinking that all the libraries in pymomentum needed to be in C++, this was mostly because I didn't quite understand how the build system handled pybinded libraries.  It was also useful at the time because the C++ libraries could call into the quaternion library, but I think this is going to be less necessary in the future.

If we rewrite the code to be Python-only, this should make future maintenance simpler and also makes the code more transparent and easier to follow.  Since it doesn't need access to any c++ momentum internals, this seems like a strictly positive change.

Reviewed By: jeongseok-meta

Differential Revision: D70038539

fbshipit-source-id: f4c1af5ad0a97069712c32b4b291a5300381e7f5