Return barycentric coordinates from closest surface point query (#253)
Summary:
Pull Request resolved: https://github.com/facebookincubator/momentum/pull/253

The closest surface point query currently returns only the projection point. However, it is sometimes useful to utilize the barycentric coordinates of the triangle primitive that are already computed in the query. This Diff aims to allow for the return of this value by adding a new field to the result struct.

The new field is declared as optional since it is not yet supported by all queries and accelerators.

Reviewed By: juliencbmeta, TuurStuyck

Differential Revision: D72196845