Add support for find_closest_points_on_mesh to pymomentum. (#254)
Summary:
Pull Request resolved: https://github.com/facebookincubator/momentum/pull/254

This is a mesh-based version of the old KdTree, which will be useful for doing surface queries.

I was considering doing this with numpy primitives instead of torch Tensors since it's not differentiable but for now I guess let's stick with torch Tensors for consistency with the existing closest_point functions, we can switch them all later if we want.

Reviewed By: jeongseok-meta

Differential Revision: D72259812

fbshipit-source-id: f9a8e03de35dcfcd8e54e1b146fcb43a2403eaff